Decoupled EncodedConnectionTest and JSONEncoderTest
